% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-foo.R
\name{foo}
\alias{foo}
\title{example function}
\usage{
foo(
  x,
  case = c("lower", "upper", "sentence"),
  spacing = 0,
  add = NULL,
  ignore = stop("STOP THIS MADNESS!")
)
}
\arguments{
\item{x}{character vector}

\item{case}{case that \code{x} will be put in: "upper", "lower" or "sentence";
can be specified in full or by partial match, e.g. the just first letter}

\item{spacing}{integer specifying character spacing in \code{x}}

\item{add}{(optional) text added before each string}

\item{ignore}{ignored argument}
}
\value{
A character string composed of \code{added} text and \code{x}.
}
\description{
A simple function to demonstrate how functions work.
}
\details{
This is a not so simple function that does a simple task:
given a character string it can change its case and character spacing,
as well as add some text in front of it.
}
\section{Purpose}{

The purpose of this function is to demonstrate how functions work.

\code{foo} has four arguments.
The first has no default and the function will throw an error if it is not supplied.
The second has three possible values that can be partially matched and the first one is the default.
The third has a default but can be set to any value, possibly generating warnings.
The fourth is optional and defaults to \code{NULL}.
The fifth is not evaluated at all.
}

\examples{
foo("bar", case = "sentence", spacing = 1)

}
