% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-grade.R
\name{grade}
\alias{grade}
\title{grade offensive graffiti}
\usage{
grade(slogan, correct = "Romani ite domum")
}
\arguments{
\item{slogan}{slogan to grade}

\item{correct}{correct form of the slogan}
}
\value{
A character vector of grades in the form of 1/3, meaning 1 out of three words are correct.
}
\description{
Check a painted slogan against its proper Latin and grade it
between 0 and 1 based on the number of the words that are correct.
}
