% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-hint.R
\name{hint}
\alias{hint}
\title{homework hints}
\usage{
hint(topic)
}
\arguments{
\item{topic}{single character string, see \code{Detials}}
}
\value{
Does not return, merely prints text to the console.
}
\description{
Get a hint for a homework assignment.
}
\details{
You can request hints for a whole course part (\code{topic = 1}),
a specific task (\code{topic = 1.1}), or ask for a specific hint (\code{topic = 1.1.1}).
}
