% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{homework}
\alias{homework}
\title{homework assignments}
\description{
This help page will contain homework assignments.
}
\details{
If you're stuck, use the \code{hint} function.
}
\section{Solutions}{

As a rule, solutions will not be provided unless I feel the exercises become too difficult or tedious.
}

\section{For Part One}{

The lecture and notes should be sufficient to solve all these problems.
When in doubt, feel free to use the help
but for the time being refrain from looking on the Internet.

Since exersises 8-11 are getting large and difficult, there is a solution provided.
Use the \code{solution} function to get a character vector like the one you're suppose to construct.
You can then test whether your result is EXACTLY like mine with \code{identical}.

\enumerate{
  \item{
  Create and print a character string describing the current date like 25th February 2020.
  }
  \item{
  Assign the previous string to a variable and print it in the console.
  }
  \item{
  Create the same string by combining three separate strings:
  one for the day, one for the month, and one for the year.
  }
  \item{
  Create three vectors that can be used to construct every date in 2000, 2010 and 2020:
  one for days (1st through 31st), one for months (January through December), and one for the years.
  }
  \item{
  Construct all dates in January 2020, each in one separate string.
  Have them printed in the console.
  }
  \item{
  Put all dates constructed above into one vector.
  }
  \item{
  Repeat the above with all dates in June 2020. Remember, there is no June 31st.
  }
  \item{
  Repeat the above with all dates in all of the three years. Mind the leap year(s).
  }
  \item{
  Repeat the above but make it so that the last date is 3rd May 2020.
  }
  \item{
  Repeat exercise 8 but make it so that state holidays are not included.
  }
  \item{
  This is a hard one.

  Repeat exercise 8 but also add week days: 25th February 2020 Tuesday.
  2000 started on a Saturday, 2010 on a Friday, and 2020 - on a Wednesday.

  Even though we did cover it, you don't need subsetting to do this, so don't.
  }
  \item{
  There are probably several objects in your Global Environment. (If not, create some.)
  Remove them all with a single console command (two at most).
  Try to do it in a way that does not require you to know the contents of GlobalEnv to do it.
  }
}
}

