% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{iris}
\alias{iris}
\title{part of Edgar Anderson's Iris Data}
\format{
An object of class \code{data.frame} with 15 rows and 5 columns.
}
\usage{
iris
}
\description{
A small subset of the built-in \code{datasets::iris} data set, containing
5 randomly selected observations from each species.
}
\keyword{datasets}
