% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{modified version of the Motor Trend Car Road Tests}
\format{
An object of class \code{data.frame} with 32 rows and 13 columns.
}
\usage{
mtcars
}
\description{
The built-in \code{datasets::mtcars} data set with added "make" and "model" columns.
}
\keyword{datasets}
