% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdlazer-package.R
\docType{package}
\name{Rdlazer-package}
\alias{Rdlazer}
\alias{Rdlazer-package}
\title{a very basic course in R}
\description{
This package contains resources used in my course in R for for extreme beginners.
}
\details{
The major feature is the Romani ite domum debacle.

There are vignettes that contain lecture notes.
Use \code{browseVignettes("Rdlazer")} to see them.

The \code{foo} function demonstrates the workings of functions.

There are also homework assignments: \code{\link[Rdlazer]{homework}}.
}
\author{
\strong{Maintainer}: Aleksander Chlebowski \email{olobiolo@gmail.com}

}
\keyword{internal}
