% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-Rdlazer.R
\docType{package}
\name{Rdlazer}
\alias{Rdlazer}
\alias{Rdlazer-package}
\title{a very basic course in R}
\description{
This package contains resources used in my course inr for for extreme beginners.
}
\details{
The major features is the Romani ite domum debacle.
}
