% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-solution.check.R
\name{solution.check}
\alias{solution.check}
\title{check your solution}
\usage{
solution.check(result, years.incl = NULL, months.incl = NULL)
}
\arguments{
\item{result}{character vector of dates}

\item{years.incl}{character vector of years to include in the table}

\item{months.incl}{character vector of months to include in the talbe}
}
\description{
Create a neat table of the dates you created.
}
\section{}{
 Warning
If you try to get a table for three years, it may not print completely.
In that case, limit it to a single year with the \code{years} argument.
}

