% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-solution.R
\name{solution}
\alias{solution}
\title{solution for dates}
\usage{
solution(skip = NULL, end = NULL, weekdays = FALSE)
}
\arguments{
\item{skip}{character vector of dates to skip}

\item{end}{character string of date to end on}

\item{weekdays}{logical flag whether to add weekdays}
}
\value{
A character vector of dates in the format "25th February 2020 Tuesday".
The week day is added depending on the \code{weekdays} argument.
Particular days can be skipped and the whole vector can end at a specific date.
}
\description{
Reproduce a result for exercises 8-11 created by me for comparison.
}
\seealso{
\code{\link[Rdlazer]{solution.check}}
}
