% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-spelling.R
\name{make_mistake}
\alias{make_mistake}
\alias{pick_position}
\title{corrupt spelling of a slogan}
\usage{
make_mistake(string)

pick_position(string, positions)
}
\arguments{
\item{string}{string to make errors in}

\item{positions}{number of positions to do typos at; taken from \code{goBrian}}
}
\value{
\code{make_mistake} returns a \code{slogan} with spelling error(s) introduced.
\code{pick_position} returns an integer vector.
}
\description{
As Brian is rather poor at Latin, he will make random spelling errors when on his mission.
}
\details{
These are internal functions used by \code{goBrian}.
}
\section{Functions}{
\itemize{
\item \code{make_mistake}: At the moment only does substitutions; removing or adding characters is not supported.
All mistakes are in lowercase at the moment.

\item \code{pick_position}: Randomly picks positions where mistakes will be made.
It runs recursively until all picked positions correspond to letters.
The first letter will always be correct.
}}

\keyword{internal}
