% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-hint.R
\name{hint}
\alias{hint}
\title{homework hints}
\usage{
hint(topic)
}
\arguments{
\item{topic}{single character string, see \code{Detials}}
}
\description{
Get a hint for a homework assignment.
}
\details{
When you request a hint, you must specify the \code{topic}.
This can be a single esercise, a specific hint for a specific esercise,
or all hints for all esercises for a specific part of the course.
In every case \code{topic} contains the index of the hint(s) you request
in the form of \code{"<part>.<esercise>.<hint>"}.

Remember, the hints are there to guide you, not constraint you.
You should use them as advice, not as imperative statements.
They assume one possible solution but most of the time more exist.
Take the hints as descriptive rather than perscriptive.
}
\section{Warning}{

I suggest you take your time before requesting a hint.
Also, be careful to get one hint at a time lest they spoil your success.
And your fun!
}

\examples{
hint('0')
hint('0.1')
hint('0.1.1')
}
