% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{homework}
\alias{homework}
\title{homework assignments}
\description{
This help page will contain homework assignments.
}
\details{
If you're stuck, use the \code{hint} function.
}
\section{For Part One}{

Use only base R when solving these problems.
\enumerate{
  \item{
  There are a number of data sets built in to R.
  They are attached to every R session and reside in the datasets package.
  Find out which of them are data frames that contain at least two columns that are factors.
  Present the result as a character vector.
  }
  \item{
  Pick 10 rows from the iris data set at random.
  }
  \item{
  Obtain a data frame like Iris, with 5 random irises from every species in the iris data set.
  }
}
}

