% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{homework}
\alias{homework}
\title{homework assignments}
\description{
This help page will contain homework assignments.
}
\details{
If you're stuck, use the \code{hint} function.
}
\section{Solutions}{

As a rule, solutions will not be provided unless I feel the exercises become too difficult or tedious.
}

\section{NA}{

As a rule, solutions will not be provided unless I feel the exercises become too difficult or tedious.
}

