% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdlazer-package.R
\docType{package}
\name{Rdlazer-package}
\alias{Rdlazer}
\alias{Rdlazer-package}
\title{a very basic course in R}
\description{
This package contains resources used in my course inr for for extreme beginners.
The major feature is the Romani ite domum debacle.
}
\details{
There are also homeworks: \code{\link[Rdlazer]{homework}}.
}
\author{
\strong{Maintainer}: Aleksander Chlebowski \email{olobiolo@gmail.com}

}
\keyword{internal}
